feat: react vite scaffold issues fixing and refactoring
