+++ lib/scaffold-react-project-structure.js
@@ -1,6 +1,7 @@
 // lib/scaffold-react-project-structure.js
 const fs = require("fs");
 const path = require("path");
+const { createMUIPages } = require("./pages/mui");
 
 function writeFileSafe(p, content) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
@@ -137,795 +138,6 @@ export const router = createBrowserRouter([
   } else {
     // Default to MUI if no specific UI is chosen
     createMUIPages(pagesDir, extx, ext, useTS);
-  }
-}
-
-function createMUIPages(pagesDir, extx, ext, useTS) {
-  const home = path.join(pagesDir, `Home.${extx}`);
-  if (!fs.existsSync(home)) {
-    writeFileSafe(home, `import { Box, Typography, Grid, Card, CardContent, Button, Avatar, Container, Paper } from '@mui/material';
-import { TrendingUp, People, Speed, Security, Code, Palette } from '@mui/icons-material';
-
-export default function Home() {
-  const features = [
-    { icon: <TrendingUp />, title: 'Performance', description: 'Lightning-fast builds with modern tooling', color: 'primary' },
-    { icon: <People />, title: 'Team Ready', description: 'Built for collaboration and scalability', color: 'secondary' },
-    { icon: <Speed />, title: 'Fast Development', description: 'Hot reload and instant feedback', color: 'success' },
-    { icon: <Security />, title: 'Type Safe', description: 'Full TypeScript support with strict mode', color: 'warning' },
-    { icon: <Code />, title: 'Clean Code', description: 'ESLint + Prettier for code quality', color: 'info' },
-    { icon: <Palette />, title: 'Beautiful UI', description: 'Material-UI components with theming', color: 'error' }
-  ];
-
-  return (
-    <Container maxWidth="lg" sx={{ py: 4 }}>
-      <Box textAlign="center" mb={6}>
-        <Typography variant="h2" component="h1" gutterBottom color="primary" fontWeight="bold">
-          Welcome to PolyFront
-        </Typography>
-        <Typography variant="h5" color="text.secondary" paragraph>
-          Your comprehensive multi-frontend scaffold with everything you need
-        </Typography>
-        <Typography variant="body1" color="text.secondary" sx={{ maxWidth: 600, mx: 'auto' }}>
-          Built with React, TypeScript, Vite, Material-UI, and modern development practices.
-          Start building your next great application today.
-        </Typography>
-      </Box>
-
-      <Grid container spacing={4} mb={6}>
-        {features.map((feature, index) => (
-          <Grid size={{ xs: 12, sm:6, md: 4 }} key={index}>
-            <Card sx={{ height: '100%', display: 'flex', flexDirection: 'column', transition: '0.3s', '&:hover': { transform: 'translateY(-4px)', boxShadow: 4 } }}>
-              <CardContent sx={{ flexGrow: 1, textAlign: 'center' }}>
-                <Avatar sx={{ bgcolor: \`\${feature.color}.main\`, width: 64, height: 64, mx: 'auto', mb: 2 }}>
-                  {feature.icon}
-                </Avatar>
-                <Typography variant="h6" component="h3" gutterBottom>
-                  {feature.title}
-                </Typography>
-                <Typography variant="body2" color="text.secondary">
-                  {feature.description}
-                </Typography>
-              </CardContent>
-            </Card>
-          </Grid>
-        ))}
-      </Grid>
-
-      <Paper elevation={2} sx={{ p: 4, textAlign: 'center' }}>
-        <Typography variant="h4" gutterBottom>
-          Ready to Get Started?
-        </Typography>
-        <Typography variant="body1" color="text.secondary" paragraph>
-          This scaffold includes routing, state management, API integration, testing, and much more.
-          Explore the different pages to see all the features in action.
-        </Typography>
-        <Box sx={{ mt: 3 }}>
-          <Button variant="contained" size="large" sx={{ mr: 2 }}>
-            View Documentation
-          </Button>
-          <Button variant="outlined" size="large">
-            Learn More
-          </Button>
-        </Box>
-      </Paper>
-    </Container>
-  );
-}
-`);
-  }
-
-  // Create other MUI pages (About, Dashboard, Profile)
-  createOtherMUIPages(pagesDir, extx, ext, useTS);
-}
-
-function createOtherMUIPages(pagesDir, extx, ext, useTS) {
-  // About page
-  const about = path.join(pagesDir, `About.${extx}`);
-  if (!fs.existsSync(about)) {
-    writeFileSafe(about, `import React from 'react';
-import { Box, Typography, Container, Paper, List, ListItem, ListItemIcon, ListItemText, Divider, Chip } from '@mui/material';
-import Grid from '@mui/material/Grid';
-import { CheckCircle, Star, Speed, Group } from '@mui/icons-material';
-
-export default function About() {
-  const techStack = [
-    { name: 'React 19', version: 'Latest', category: 'Framework' },
-    { name: 'TypeScript', version: '5.x', category: 'Language' },
-    { name: 'Vite', version: '7.x', category: 'Build Tool' },
-    { name: 'Material-UI', version: '7.x', category: 'UI Library' },
-    { name: 'React Router', version: '7.x', category: 'Routing' },
-    { name: 'Zustand', version: '5.x', category: 'State Management' },
-    { name: 'Axios', version: '1.x', category: 'HTTP Client' },
-    { name: 'Vitest', version: '3.x', category: 'Testing' }
-  ];
-
-  const features = [
-    'Modern React with hooks and functional components',
-    'Full TypeScript support with strict mode',
-    'Beautiful Material-UI components with theming',
-    'Client-side routing with React Router',
-    'State management with Zustand',
-    'API integration with Axios',
-    'Comprehensive testing setup with Vitest',
-    'ESLint and Prettier for code quality',
-    'Hot module replacement for fast development',
-    'Responsive design with mobile-first approach'
-  ];
-
-  return (
-    <Container maxWidth="lg" sx={{ py: 4 }}>
-      <Box textAlign="center"  mb={6}>
-        <Typography variant="h2" component="h1" gutterBottom color="primary" fontWeight="bold">
-          About PolyFront
-        </Typography>
-        <Typography variant="h5" color="text.secondary" paragraph>
-          A comprehensive scaffold for modern web applications
-        </Typography>
-      </Box>
-
-      <Grid container spacing={4} mb={6}>
-        <Grid size={{ xs: 12, md: 6 }}>
-          <Paper elevation={2} sx={{ p: 4, height: '100%' }}>
-            <Typography variant="h4" gutterBottom color="primary">
-              Technology Stack
-            </Typography>
-            <Typography variant="body1" paragraph>
-              Built with the latest and greatest technologies in the React ecosystem.
-            </Typography>
-            <Grid container spacing={2}>
-              {techStack.map((tech, index) => (
-                <Grid size={{ xs: 6 }} key={index}>
-                  <Chip label={\`\${tech.name} \${tech.version}\`} variant="outlined" color="primary" size="small" sx={{ mb: 1 }} />
-                  <Typography variant="caption" display="block" color="text.secondary">{tech.category}</Typography>
-                </Grid>
-              ))}
-            </Grid>
-          </Paper>
-        </Grid>
-
-        <Grid size={{ xs: 12, md: 6 }}>
-          <Paper elevation={2} sx={{ p: 4, height: '100%' }}>
-            <Typography variant="h4" gutterBottom color="primary">Key Features</Typography>
-            <List>
-              {features.map((feature, index) => (
-                <React.Fragment key={index}>
-                  <ListItem>
-                    <ListItemIcon><CheckCircle color="success" /></ListItemIcon>
-                    <ListItemText primary={feature} />
-                  </ListItem>
-                  {index < features.length - 1 && <Divider />}
-                </React.Fragment>
-              ))}
-            </List>
-          </Paper>
-        </Grid>
-      </Grid>
-
-      <Paper elevation={2} sx={{ p: 4, textAlign: 'center' }}>
-        <Typography variant="h4" gutterBottom>Why Choose PolyFront?</Typography>
-        <Grid container spacing={3} sx={{ mt: 2 }}>
-          <Grid size={{ xs: 12, sm: 4 }}>
-            <Star color="primary" sx={{ fontSize: 40, mb: 1 }} />
-            <Typography variant="h6" gutterBottom>Production Ready</Typography>
-            <Typography variant="body2" color="text.secondary">Built with best practices and industry standards</Typography>
-          </Grid>
-          <Grid size={{ xs: 12, sm: 4 }}>
-            <Speed color="primary" sx={{ fontSize: 40, mb: 1 }} />
-            <Typography variant="h6" gutterBottom>Fast Development</Typography>
-            <Typography variant="body2" color="text.secondary">Hot reload and instant feedback for rapid iteration</Typography>
-          </Grid>
-          <Grid size={{ xs: 12, sm: 4 }}>
-            <Group color="primary" sx={{ fontSize: 40, mb: 1 }} />
-            <Typography variant="h6" gutterBottom>Team Friendly</Typography>
-            <Typography variant="body2" color="text.secondary">Consistent code style and comprehensive tooling</Typography>
-          </Grid>
-        </Grid>
-      </Paper>
-    </Container>
-  );
-}
-`);
-
-  }
-
-  // Dashboard page
-  const dashboard = path.join(pagesDir, `Dashboard.${extx}`);
-  if (!fs.existsSync(dashboard)) {
-    writeFileSafe(dashboard, `import React,{ useState } from 'react';
-import {
-  Box,
-  Typography,
-  Card,
-  CardContent,
-  CardHeader,
-  Chip,
-  Avatar,
-  List,
-  ListItem,
-  ListItemAvatar,
-  ListItemText,
-  Button,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow
-} from '@mui/material';
-import Grid from '@mui/material/Grid';
-import {
-  TrendingUp,
-  People,
-  AttachMoney,
-  ShoppingCart,
-  Visibility,
-  CheckCircle,
-  Warning,
-  Error
-} from '@mui/icons-material';
-
-export default function Dashboard() {
-  const [stats] = useState({
-    users: 1234,
-    revenue: 45678,
-    orders: 89,
-    views: 12345
-  });
-
-  const [recentActivity] = useState([
-    { id: 1, user: 'Nirmal Samaranayaka', action: 'Order placed', time: '2 min ago', status: 'success' },
-    { id: 2, user: 'Jane Smith', action: 'Payment received', time: '5 min ago', status: 'success' },
-    { id: 3, user: 'Bob Johnson', action: 'Login attempt', time: '10 min ago', status: 'warning' },
-    { id: 4, user: 'Alice Brown', action: 'Account created', time: '15 min ago', status: 'success' },
-  ]);
-
-  const [topProducts] = useState([
-    { name: 'Premium Widget', sales: 156, revenue: 2340, growth: 12 },
-    { name: 'Super Gadget', sales: 98, revenue: 1890, growth: -5 },
-    { name: 'Mega Tool', sales: 87, revenue: 1450, growth: 8 },
-    { name: 'Ultra Device', sales: 76, revenue: 1230, growth: 15 },
-  ]);
-
-  const getStatusIcon = (status ${useTS ? ':string' : ''}) => {
-    switch (status) {
-      case 'success': return <CheckCircle color="success" />;
-      case 'warning': return <Warning color="warning" />;
-      case 'error': return <Error color="error" />;
-      default: return <CheckCircle color="success" />;
-    }
-  };
-
-  const getGrowthColor = (growth ${useTS ? ': number' : ''}) => growth >= 0 ? 'success' : 'error';
-
-  return (
-    <Box>
-      <Typography variant="h3" component="h1" gutterBottom color="primary" fontWeight="bold">
-        Dashboard
-      </Typography>
-      <Typography variant="h6" color="text.secondary" paragraph>
-        Welcome back! Here's what's happening with your application today.
-      </Typography>
-
-      {/* Stats Cards */}
-      <Grid container spacing={3} mb={4}>
-        <Grid size={{ xs: 12, sm:6, md: 3 }}>
-          <Card>
-            <CardContent>
-              <Box display="flex" alignItems="center" justifyContent="space-between">
-                <Box>
-                  <Typography color="text.secondary" gutterBottom variant="h6">
-                    Total Users
-                  </Typography>
-                  <Typography variant="h4" component="div">
-                    {stats.users.toLocaleString()}
-                  </Typography>
-                  <Typography variant="body2" color="success.main" display="flex" alignItems="center">
-                    <TrendingUp fontSize="small" sx={{ mr: 0.5 }} />
-                    +12% from last month
-                  </Typography>
-                </Box>
-                <Avatar sx={{ bgcolor: 'primary.main', width: 56, height: 56 }}>
-                  <People />
-                </Avatar>
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid>
-
-        <Grid size={{ xs: 12, sm:6, md: 3 }}>
-          <Card>
-            <CardContent>
-              <Box display="flex" alignItems="center" justifyContent="space-between">
-                <Box>
-                  <Typography color="text.secondary" gutterBottom variant="h6">
-                    Revenue
-                  </Typography>
-                  <Typography variant="h4" component="div">
-                    \${stats.revenue.toLocaleString()}
-                  </Typography>
-                  <Typography variant="body2" color="success.main" display="flex" alignItems="center">
-                    <TrendingUp fontSize="small" sx={{ mr: 0.5 }} />
-                    +8% from last month
-                  </Typography>
-                </Box>
-                <Avatar sx={{ bgcolor: 'success.main', width: 56, height: 56 }}>
-                  <AttachMoney />
-                </Avatar>
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid>
-
-        <Grid size={{ xs: 12, sm:6, md: 3 }}>
-          <Card>
-            <CardContent>
-              <Box display="flex" alignItems="center" justifyContent="space-between">
-                <Box>
-                  <Typography color="text.secondary" gutterBottom variant="h6">
-                    Orders
-                  </Typography>
-                  <Typography variant="h4" component="div">
-                    {stats.orders}
-                  </Typography>
-                  <Typography variant="body2" color="success.main" display="flex" alignItems="center">
-                    <TrendingUp fontSize="small" sx={{ mr: 0.5 }} />
-                    +15% from last month
-                  </Typography>
-                </Box>
-                <Avatar sx={{ bgcolor: 'warning.main', width: 56, height: 56 }}>
-                  <ShoppingCart />
-                </Avatar>
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid>
-
-        <Grid size={{ xs: 12, sm:6, md: 3 }}>
-          <Card>
-            <CardContent>
-              <Box display="flex" alignItems="center" justifyContent="space-between">
-                <Box>
-                  <Typography color="text.secondary" gutterBottom variant="h6">
-                    Page Views
-                  </Typography>
-                  <Typography variant="h4" component="div">
-                    {stats.views.toLocaleString()}
-                  </Typography>
-                  <Typography variant="body2" color="success.main" display="flex" alignItems="center">
-                    <TrendingUp fontSize="small" sx={{ mr: 0.5 }} />
-                    +22% from last month
-                  </Typography>
-                </Box>
-                <Avatar sx={{ bgcolor: 'info.main', width: 56, height: 56 }}>
-                  <Visibility />
-                </Avatar>
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid>
-      </Grid>
-
-      {/* Charts and Tables */}
-      <Grid container spacing={3}>
-        <Grid size={{ xs: 12, md: 8 }}>
-          <Card>
-            <CardHeader
-              title="Top Products"
-              action={
-                <Button size="small" color="primary">
-                  View All
-                </Button>
-              }
-            />
-            <CardContent>
-              <TableContainer>
-                <Table>
-                  <TableHead>
-                    <TableRow>
-                      <TableCell>Product</TableCell>
-                      <TableCell align="right">Sales</TableCell>
-                      <TableCell align="right">Revenue</TableCell>
-                      <TableCell align="right">Growth</TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {topProducts.map((product) => (
-                      <TableRow key={product.name}>
-                        <TableCell component="th" scope="row">
-                          {product.name}
-                        </TableCell>
-                        <TableCell align="right">{product.sales}</TableCell>
-                        <TableCell align="right">\${product.revenue.toLocaleString()}</TableCell>
-                        <TableCell align="right">
-                          <Chip
-                            label={\`\${product.growth >= 0 ? '+' : ''}\${product.growth}%\`}
-                            color={getGrowthColor(product.growth)}
-                            size="small"
-                          />
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </TableContainer>
-            </CardContent>
-          </Card>
-        </Grid>
-
-        <Grid size={{ xs: 12, md: 4 }}>
-          <Card>
-            <CardHeader title="Recent Activity" />
-            <CardContent>
-              <List>
-                {recentActivity.map((activity) => (
-                  <ListItem key={activity.id} alignItems="flex-start">
-                    <ListItemAvatar>
-                      <Avatar sx={{ bgcolor: 'grey.100' }}>
-                        {getStatusIcon(activity.status)}
-                      </Avatar>
-                    </ListItemAvatar>
-                    <ListItemText
-                      primary={activity.action}
-                      secondary={
-                        <React.Fragment>
-                          <Typography component="span" variant="body2" color="text.primary">
-                            {activity.user}
-                        </Typography>
-                          {' — '}{activity.time}
-                        </React.Fragment>
-                      }
-                    />
-                  </ListItem>
-                ))}
-              </List>
-            </CardContent>
-          </Card>
-        </Grid>
-      </Grid>
-    </Box>
-  );
-}`);
-  }
-
-  // Profile page
-  const profile = path.join(pagesDir, `Profile.${extx}`);
-  if (!fs.existsSync(profile)) {
-    writeFileSafe(profile, `import { useState } from 'react';
-import {
-  Box,
-  Typography,
-  Grid,
-  Card,
-  CardContent,
-  CardHeader,
-  Avatar,
-  Button,
-  TextField,
-  Chip,
-  List,
-  ListItem,
-  ListItemText,
-  ListItemIcon,
-  Switch
-} from '@mui/material';
-import {
-  Edit,
-  Save,
-  Cancel,
-  Email,
-  Phone,
-  LocationOn,
-  Work,
-  Notifications,
-  Security,
-  Palette
-} from '@mui/icons-material';
-
-export default function Profile() {
-  const [isEditing, setIsEditing] = useState(false);
-  const [profile] = useState({
-    name: 'Nirmal Samaranayaka',
-    email: 'nirmal.fullstack@gmail.com',
-    phone: '+46 (72) xxx-xxxx',
-    location: 'Stockholm, Sweden',
-    company: 'Scania AB.',
-    position: 'Senior Fullstack Developer',
-    education: 'Computer Science, University of Colombo',
-    website: 'https://dev.to/nirmalsamaranayaka',
-    bio: 'Experienced Full Stack Engineer & Tech Lead | Specialized in .NET, React, Angular, and scalable cloud-native solutions.'
-  });
-
-  const [settings, setSettings] = useState({
-    emailNotifications: true,
-    pushNotifications: false,
-    darkMode: false,
-    twoFactorAuth: true
-  });
-
- const [skills] = useState([
-  // Frontend
-  'React', 'Redux', 'MobX', 'React Query', 'Angular', 'Micro Frontends', 
-  'TypeScript', 'JavaScript', 'HTML/CSS', 'jQuery',
-
-  // UI Libraries & Styling
-  'MUI (Material-UI)', 'Kendo UI', 'Tailwind CSS', 'Tegel',
-
-  // Backend
-  '.NET 6/7/8', 'ASP.NET Core', 'C#', 'Web API', 'WCF', 
-  'Microservices', 'gRPC', 'REST APIs', 'SOAP', 'Node.js',
-
-  // Blockchain
-  'Solidity', 'Truffle',
-
-  // Cloud & DevOps
-  'Azure', 'AWS', 'Docker', 'Kubernetes', 'Helm', 
-  'CI/CD', 'Azure DevOps', 'Jenkins', 'Git', 'TFS/TFVC',
-
-  // Data & BI
-  'MSSQL', 'Oracle', 'Entity Framework Core', 'Entity Framework 6', 'Dapper', 
-  'T-SQL', 'PL-SQL', 'SSIS', 'SSRS', 'Data Migration', 'Query Tuning',
-
-  // Architecture & Patterns
-  'Clean Architecture', 'Domain-Driven Design (DDD)', 'SOLID', 
-  'Object-Oriented Programming (OOP)', 'Test-Driven Development (TDD)', 
-  'MVVM', 'Agile (Scrum, Kanban)'
-]);
-
-  const handleSave = () => {
-    setIsEditing(false);
-    // Here you would typically save to an API
-  };
-
-  const handleCancel = () => {
-    setIsEditing(false);
-    // Reset to original values
-  };
-
-  const handleSettingChange = (setting  ${useTS ? ':string' : ''}) => (event ${useTS ? ': React.ChangeEvent<HTMLInputElement>' : ''} ) => {
-    setSettings(prev => ({
-      ...prev,
-      [setting]: event.target.checked
-    }));
-  };
-
-  return (
-    <Box>
-      <Box display="flex" justifyContent="space-between" alignItems="center" mb={3}>
-        <Typography variant="h3" component="h1" color="primary" fontWeight="bold">
-          Profile
-        </Typography>
-        <Box>
-          {isEditing ? (
-            <>
-              <Button
-                variant="contained"
-                startIcon={<Save />}
-                onClick={handleSave}
-                sx={{ mr: 1 }}
-              >
-                Save
-              </Button>
-              <Button
-                variant="outlined"
-                startIcon={<Cancel />}
-                onClick={handleCancel}
-              >
-                Cancel
-              </Button>
-            </>
-          ) : (
-            <Button
-              variant="contained"
-              startIcon={<Edit />}
-              onClick={() => setIsEditing(true)}
-            >
-              Edit Profile
-            </Button>
-          )}
-        </Box>
-      </Box>
-
-      <Grid container spacing={3}>
-        {/* Profile Header */}
-        <Grid size={{ xs: 12}}>
-          <Card>
-            <CardContent>
-              <Box display="flex" alignItems="center" gap={3}>
-                <Avatar
-                  sx={{ width: 120, height: 120, fontSize: '3rem' }}
-                  src="/static/images/avatar/1.jpg"
-                >
-                  {profile.name.split(' ').map(n => n[0]).join('')}
-                </Avatar>
-                <Box flex={1}>
-                  <Typography variant="h4" gutterBottom>
-                    {profile.name}
-                  </Typography>
-                  <Typography variant="h6" color="text.secondary" gutterBottom>
-                    {profile.position} at {profile.company}
-                  </Typography>
-                  <Typography variant="body1" color="text.secondary" paragraph>
-                    {profile.bio}
-                  </Typography>
-                  <Box display="flex" gap={1} flexWrap="wrap">
-                    {skills.map((skill) => (
-                      <Chip key={skill} label={skill} variant="outlined" size="small" />
-                    ))}
-                  </Box>
-                </Box>
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid>
-
-        {/* Profile Details */}
-        <Grid size={{ xs: 12, md: 8 }}>
-          <Card>
-            <CardHeader title="Personal Information" />
-            <CardContent>
-              <Grid container spacing={3}>
-                <Grid size={{ xs: 12, sm: 6 }}>
-                  <TextField
-                    fullWidth
-                    label="Full Name"
-                    value={profile.name}
-                    disabled={!isEditing}
-                    InputProps={{
-                      startAdornment: <Email sx={{ mr: 1, color: 'text.secondary' }} />
-                    }}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, sm: 6 }}>
-                  <TextField
-                    fullWidth
-                    label="Email"
-                    value={profile.email}
-                    disabled={!isEditing}
-                    InputProps={{
-                      startAdornment: <Email sx={{ mr: 1, color: 'text.secondary' }} />
-                    }}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, sm: 6 }}>
-                  <TextField
-                    fullWidth
-                    label="Phone"
-                    value={profile.phone}
-                    disabled={!isEditing}
-                    InputProps={{
-                      startAdornment: <Phone sx={{ mr: 1, color: 'text.secondary' }} />
-                    }}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, sm: 6 }}>
-                  <TextField
-                    fullWidth
-                    label="Location"
-                    value={profile.location}
-                    disabled={!isEditing}
-                    InputProps={{
-                      startAdornment: <LocationOn sx={{ mr: 1, color: 'text.secondary' }} />
-                    }}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, sm: 6 }}>
-                  <TextField
-                    fullWidth
-                    label="Company"
-                    value={profile.company}
-                    disabled={!isEditing}
-                    InputProps={{
-                      startAdornment: <Work sx={{ mr: 1, color: 'text.secondary' }} />
-                    }}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12, sm: 6 }}>
-                  <TextField
-                    fullWidth
-                    label="Position"
-                    value={profile.position}
-                    disabled={!isEditing}
-                    InputProps={{
-                      startAdornment: <Work sx={{ mr: 1, color: 'text.secondary' }} />
-                    }}
-                  />
-                </Grid>
-                <Grid size={{ xs: 12 }}>
-                  <TextField
-                    fullWidth
-                    multiline
-                    rows={3}
-                    label="Bio"
-                    value={profile.bio}
-                    disabled={!isEditing}
-                  />
-                </Grid>
-              </Grid>
-            </CardContent>
-          </Card>
-        </Grid>
-
-        {/* Settings */}
-        <Grid size={{ xs: 12, md: 4 }}>
-          <Card>
-            <CardHeader title="Settings" />
-            <CardContent>
-              <List>
-                <ListItem>
-                  <ListItemIcon>
-                    <Notifications />
-                  </ListItemIcon>
-                  <ListItemText primary="Email Notifications" />
-                  <Switch
-                    checked={settings.emailNotifications}
-                    onChange={handleSettingChange('emailNotifications')}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemIcon>
-                    <Notifications />
-                  </ListItemIcon>
-                  <ListItemText primary="Push Notifications" />
-                  <Switch
-                    checked={settings.pushNotifications}
-                    onChange={handleSettingChange('pushNotifications')}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemIcon>
-                    <Palette />
-                  </ListItemIcon>
-                  <ListItemText primary="Dark Mode" />
-                  <Switch
-                    checked={settings.darkMode}
-                    onChange={handleSettingChange('darkMode')}
-                  />
-                </ListItem>
-                <ListItem>
-                  <ListItemIcon>
-                    <Security />
-                  </ListItemIcon>
-                  <ListItemText primary="Two-Factor Auth" />
-                  <Switch
-                    checked={settings.twoFactorAuth}
-                    onChange={handleSettingChange('twoFactorAuth')}
-                  />
-                </ListItem>
-              </List>
-            </CardContent>
-          </Card>
-
-          <Card sx={{ mt: 2 }}>
-            <CardHeader title="Quick Actions" />
-            <CardContent>
-              <Button
-                fullWidth
-                variant="outlined"
-                sx={{ mb: 1 }}
-              >
-                Download Resume
-              </Button>
-              <Button
-                fullWidth
-                variant="outlined"
-                sx={{ mb: 1 }}
-              >
-                Share Profile
-              </Button>
-              <Button
-                fullWidth
-                variant="outlined"
-                color="error"
-              >
-                Delete Account
-              </Button>
-            </CardContent>
-          </Card>
-        </Grid>
-      </Grid>
-    </Box>
-  );
-}`);
   }
 }
 
@@ -4360,7 +3572,7 @@ const Layout = () => {
           </Typography>
 
           <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-            <Chip label="v1.0.22" size="small" variant="outlined" />
+            <Chip label="v1.0.23" size="small" variant="outlined" />
             {/* Language selector using i18n */}
             <Select
               value={currentLang}
@@ -4505,7 +3717,7 @@ const Layout = () => {
             </button>
             <h4 className="mb-0">PolyFront</h4>
             <div className="d-flex align-items-center gap-2">
-              <span className="badge bg-secondary">v1.0.22</span>
+              <span className="badge bg-secondary">v1.0.23</span>
               
               {/* Language Selector */}
               <select
@@ -4599,7 +3811,7 @@ export default Layout;`);
                 {/* Version + Language + User */}
                 <div className="flex items-center gap-3">
                   {/* Version */}
-                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.22</span>
+                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.23</span>
       
                   {/* Language selector */}
                   <select
@@ -4708,7 +3920,7 @@ style={{ borderRight: 0 }}
         >
           PolyFront
         </Typography.Title>
-        <Tag bordered={false}>v1.0.22</Tag>
+        <Tag bordered={false}>v1.0.23</Tag>
         <Select
           size="small"
           value={currentLang}
@@ -4902,7 +4114,7 @@ function Layout() {
           <Box flex="1" />
 
           <HStack gap={2}>
-            <Badge>v1.0.22</Badge>
+            <Badge>v1.0.23</Badge>
 
             {/* Language Select (v3 controlled) */}
             <Select.Root
